Added Ultra Greed teleport Add debug variable GreedSpecialRooms.SacrificeMin to lower the minimum required sacrifices to start rolling for a teleport Convert remaining prints to debugPrint Turn debug off Delete old main I think this is ready for release
