+++ scripts/getroomdata.lua
@@ -3,6 +3,7 @@ local game = Game()
 local rng = mod.rng
 
 mod.roomdata = {}
+mod.dotransition = false
 
 local stringToType = {
 	["planetarium"] = RoomType.ROOM_PLANETARIUM,
@@ -13,6 +14,7 @@ local function GetRoomData(key, curse)
 	mod.lib.debugPrint("GetRoomData.. key: "..key.." curse: "..tostring(curse))
 	local level = game:GetLevel()
 	Isaac.ExecuteCommand("goto s." .. mod.SpecialRoom[key].string .. "." .. mod.GetRoomID(key, rng))
+	mod.dotransition = true --we are bound to it the moment we change rooms from the start
 	local gotor = level:GetRoomByIdx(-3,0)
 	if gotor.Data then
 		if curse then
@@ -37,8 +39,8 @@ function mod.GetCustomRoomData()
 	end
 	mod.redRoomsRequired = 0
 	for i,v in pairs(mod.roomsrequested.redRoom) do
-		print(i.." "..tostring(v))
-		print(stringToType[i])
+		mod.lib.debugPrint(i.." "..tostring(v))
+		mod.lib.debugPrint(stringToType[i])
 		if v > RoomType.ROOM_NULL and GetRoomData(stringToType[i], false) then
 			mod.redRoomsRequired = mod.redRoomsRequired + 1
 		end
