+++ scripts/pause.lua
@@ -1,0 +1,42 @@
+local mod = GreedSpecialRooms
+local game = Game()
+
+mod.paused = false
+
+--player:AddControlsCooldown(int) could work too, but we want to mimic the whole game pausing
+function mod:PauseGame(force)
+	if game:GetRoom():GetBossID() ~= 54 or force then -- Intentionally fail achievement note pauses on Lamb, since it breaks the Victory Lap menu super hard
+		for _, projectile in pairs(Isaac.FindByType(9)) do
+			projectile:Remove()
+
+			local poof = Isaac.Spawn(1000, 15, 0, projectile.Position, Vector.Zero, nil)
+			poof.SpriteScale = Vector.One * 0.75
+		end
+
+		for _, pillar in pairs(Isaac.FindByType(951, 1)) do
+			pillar:Kill()
+			pillar:Remove()
+		end
+
+		mod.paused = true
+
+		Isaac.GetPlayer():UseActiveItem(CollectibleType.COLLECTIBLE_PAUSE, UseFlag.USE_NOANIM)
+	end
+end
+
+for hook = InputHook.IS_ACTION_PRESSED, InputHook.IS_ACTION_TRIGGERED do
+	mod:AddCallback(ModCallbacks.MC_INPUT_ACTION, function(_, entity, hook, action)
+		if mod.paused and action ~= ButtonAction.ACTION_CONSOLE then
+			return false
+		end
+	end, hook)
+end
+
+mod:AddCallback(ModCallbacks.MC_INPUT_ACTION, function(_, entity, hook, action)
+	if mod.paused and action ~= ButtonAction.ACTION_CONSOLE then
+		return 0
+	end
+end, InputHook.GET_ACTION_VALUE)
+
+--just in case
+mod:AddCallback(ModCallbacks.MC_POST_GAME_STARTED, function() mod.paused = false end)
